# Improve OSX benchmark performance