We've known for a while that OSX has a slow mutex implementation.
The hot path of deque includes acquiring at least one mutex, maybe
two when sending. This meant that OSX suffered from very slow
benchmark performance compared to std MPSC, 100ms vs 1.5ms. This
commit introduces parking_lot Mutex/Condvar to close that gap. No
measurement has been done on Linux as of this writing. The benchmarks
have also been fiddled with to improve their output: we can now
compare the throughput of our benchmarks where before we had to
calculate that.

An example:

```
     Running target/release/deps/stdlib_comparison-486ac007903caae5
hopper/all in-memory/HopperInput { in_memory_max: 8192, on_disk_max: 32768, t...
                        time:   [2.3519 ms 2.3955 ms 2.4444 ms]
                        thrpt:  [3.3513 Melem/s 3.4197 Melem/s 3.4831 Melem/s]
                        change: [-1.3650% +0.5431% +2.5115%] (p = 0.60 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
hopper/all in-memory/HopperInput { in_memory_max: 4096, on_disk_max: 32768, t...
                        time:   [2.4669 ms 2.5005 ms 2.5368 ms]
                        thrpt:  [3.2293 Melem/s 3.2761 Melem/s 3.3207 Melem/s]
                        change: [-15.043% -10.065% -5.1173%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

mpsc/all in-memory      time:   [1.3947 ms 1.4059 ms 1.4209 ms]
                        thrpt:  [5.7654 Melem/s 5.8269 Melem/s 5.8738 Melem/s]
                        change: [-2.7682% -0.5352% +1.4363%] (p = 0.64 > 0.05)
                        No change in performance detected.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) high mild
  12 (12.00%) high severe
```

Linux performance should be checked before merging this PR.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>